+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v2.6.0](https://github.com/edtr-io/edtr-io/compare/v2.5.2..v2.6.0) - May 23, 2022
+
+### Added
+
+- **plugin-rows**. Add `isEmptyRows` selector.
+
 ## [v2.5.2](https://github.com/edtr-io/edtr-io/compare/v2.5.1..v2.5.2) - May 9, 2022
 
 ### Added
 
-- **plugin-image**. add caption of images
+- **plugin-image**. Add caption of images.
 
 ## [v2.5.1](https://github.com/edtr-io/edtr-io/compare/v2.5.0..v2.5.1) - April 15, 2022